Add files via upload
This project describes the design of a myolectric prosthetic hand that reads muscle signals from the emg sensor and map these to servo angles allowing the fingers to move through the motor action. There is also a touch sensing function that reads inputs from the ESP32 touch pins and gives feedback through a vibration motor. A power button switches the hand on and off, and a mode button switches it between two gesture modes.